@login_required and superuser restrictions added

+++ products/views.py
@@ -1,6 +1,8 @@
 from django.shortcuts import (
     render, redirect, reverse, get_object_or_404)
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+
 from django.db.models import Q
 from django.db.models.functions import Lower
 
@@ -118,8 +120,14 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     # add product to the store
+
+    if not request.user.is_superuser:
+        messages.error(
+            request, 'Sorry, restricted to admin privilege!')
+        return redirect(reverse('home'))
 
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
@@ -143,8 +151,14 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     # edit existing product
+
+    if not request.user.is_superuser:
+        messages.error(
+            request, 'Sorry, restricted to admin privilege!')
+        return redirect(reverse('home'))
 
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
@@ -168,8 +182,14 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     # delete product from store
+
+    if not request.user.is_superuser:
+        messages.error(
+            request, 'Sorry, restricted to admin privilege!')
+        return redirect(reverse('home'))
 
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
